Refactor: Change return types from int to int8_t and int32_t in various functions
- Updated return types of several functions in highlight.c, input.c, render.c, rows.c, and terminal.c from int to int8_t or int32_t for better type safety and clarity.
- Adjusted function parameters to use int32_t where appropriate, enhancing consistency across the codebase.
- Modified handling of cursor and row operations to accommodate the new types, ensuring compatibility with existing logic.
- Improved readability and maintainability by standardizing integer types used in the editor's core functionalities.

+++ include/buffer.h
@@ -6,13 +6,13 @@
 
 struct ABuf {
     char *buf;
-    size_t len;
+    int32_t len;
 };
 
 #define ABUF_INIT \
     { NULL, 0 }
 
-uint8_t ab_append(struct ABuf *ab, const char *s, size_t len);
-uint8_t ab_free(struct ABuf *ab);
+int8_t ab_append(struct ABuf *ab, const char *s, int32_t len);
+int8_t ab_free(struct ABuf *ab);
 
 #endif
